Additional unit tests for Event Service

+++ BoardGameHub-Backend/src/main/java/ca/mcgill/ecse321/boardgamehub/repo/RegistrationRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface RegistrationRepository extends CrudRepository<Registration, Registration.Key> {
     public Registration findRegistrationByKey(Registration.Key key);
-    long countByKey_RegisteredEvent(Event registeredEvent);
+    int countByKey_RegisteredEvent(Event registeredEvent);
 }
